feat: Add support for encoding null values in JSON:API objects
This change adds support for encoding null values in JSON:API objects. Previously, when attempting to encode a null value, an error would occur. With this enhancement, the `encode` method in the `Japx` class now properly handles null values by setting the `data` field to null in the resulting JSON:API object.

This change allows for more flexibility in the encoding process, as it enables the representation of null values in JSON:API objects. It ensures consistency with the JSON:API specification and provides developers with the ability to handle null values appropriately when encoding data.

Fixes: #37